Extract error types to errors.go
